+++ buession-security-pac4j/src/main/java/com/buession/security/pac4j/annotation/PrincipalAnnotationUtils.java
@@ -19,7 +19,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2023 Buession.com Inc.														       |
+ * | Copyright @ 2013-2026 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.security.pac4j.annotation;
@@ -49,7 +49,7 @@ public class PrincipalAnnotationUtils {
 	 * @since 2.3.2
 	 */
 	public static boolean supportsParameter(MethodParameter parameter) {
-		if(parameter.hasParameterAnnotation(Principal.class) == true){
+		if(parameter.hasParameterAnnotation(Principal.class)){
 			final Class<?> parameterType = parameter.getParameterType();
 
 			return parameterType.isPrimitive() == false && parameterType.isArray() == false &&