+++ buession-security-web/src/main/java/com/buession/security/web/config/PermissionsPolicy.java
@@ -34,26 +34,40 @@ import java.util.StringJoiner;
  */
 public class PermissionsPolicy extends BaseConfig {
 
+	/**
+	 * 策略
+	 */
 	private String policy;
 
+	/**
+	 * 构造函数
+	 */
 	public PermissionsPolicy() {
 		super(false);
 	}
 
+	/**
+	 * 返回策略
+	 *
+	 * @return 策略
+	 */
 	public String getPolicy() {
 		return policy;
 	}
 
+	/**
+	 * 设置策略
+	 *
+	 * @param policy
+	 * 		策略
+	 */
 	public void setPolicy(String policy) {
 		this.policy = policy;
 	}
 
 	@Override
 	public String toString() {
-		return new StringJoiner(", ", "{", "}")
-				.add("enabled=" + getEnabled())
-				.add("policy=" + policy)
-				.toString();
+		return new StringJoiner(", ", "{", "}").add("enabled=" + getEnabled()).add("policy=" + policy).toString();
 	}
 
 }